% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATH4753Falast2025-package.R
\docType{package}
\name{MATH4753Falast2025-package}
\alias{MATH4753Falast2025}
\alias{MATH4753Falast2025-package}
\title{MATH4753Falast2025: Math 4753 Lab Functions}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\keyword{internal}
