% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5.R
\name{mybin}
\alias{mybin}
\title{Simulates a Binomial Experiment}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations (simulations).}

\item{n}{Number of trials in each simulation.}

\item{p}{Probability of success in each trial.}
}
\value{
A table with the proportion of each number of successes.
}
\description{
This function simulates a binomial experiment by generating random samples
from a binomial distribution and visualizing the results.
}
\examples{
mybin(1000, 10, 0.7)

}
