% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab4function.R
\name{myf}
\alias{myf}
\title{A Custom Function for Piecewise Quadratic Regression}
\usage{
myf(x, coef)
}
\arguments{
\item{x}{A numeric value or vector. Represents the independent variable(s).}

\item{coef}{A numeric vector of coefficients. Must contain three elements:
- coef[1]: The intercept.
- coef[2]: The linear term coefficient.
- coef[3]: The quadratic term coefficient for (x - 18), applied only when (x > 18).}
}
\value{
A numeric value or vector representing the result of the piecewise equation.
}
\description{
This function evaluates a piecewise quadratic regression equation.
}
\examples{
myf(20, c(2, 3, 4))
myf(c(15, 18, 20), c(2, 3, 4))
}
