% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Calculate Optimal Number of Tickets to Sell without Overbooking}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Integer. Number of seats on the plane.}

\item{gamma}{Numeric between 0 and 1. Maximum acceptable probability of overbooking.}

\item{p}{Numeric between 0 and 1. Probability that a ticketed passenger shows up.}
}
\value{
A named list containing:
\describe{
  \item{nd}{Optimal number of tickets}
  \item{nc}{Optimal number of tickets}
  \item{N}{Number of seats}
  \item{p}{Probability of people showing up}
  \item{gamma}{Input overbooking threshold}
}
}
\description{
This function helps calculate the optimal number of tickets to sell for a flight without overbooking
seats, given the number of seats `N`, probability `p` of a customer showing up, and a maximum
allowed overbooking probability `gamma`. Two plost are also generated for the discrete case, and continuous case.
}
\examples{
ntickets(N = 200, gamma = 0.02, p = 0.95)

}
